Add: Separate module for `quantum_info` Mimicking what was done in the main qiskit repository, this commit separates the `Observable`, `BitTerm` and `BitTermList` into its own crate called `quantum_info`.
